entity created and fixes
Added birth date, gender, and residence fields.

Renamed the enum Active to State.

In TicketMachine, added a "type" field and set the identificationCode to be unique.

Added surname to the Seller entity.

Created DAO classes for TicketSeller and TicketMachine.

Renamed the save and findById methods in VehiclesDAO and VehiclesMaintenanceDAO to keep naming consistent across the project.

+++ src/main/java/team1/entities/VehicleLineJourney.java
@@ -10,7 +10,7 @@ public class VehicleLineJourney {
     @Id
     @GeneratedValue
     private long id;
-    private double actualTime;
+    private int actualTime;
 
     @ManyToOne
     @JoinColumn(name = "vehicles_id")
@@ -20,7 +20,7 @@ public class VehicleLineJourney {
     private Line lineId;
 
     public VehicleLineJourney(){}
-    public VehicleLineJourney( double actualTime, Vehicles vehiclesId, Line lineId) {
+    public VehicleLineJourney( int actualTime, Vehicles vehiclesId, Line lineId) {
         this.actualTime = actualTime;
         this.vehiclesId = vehiclesId;
         this.lineId = lineId;
@@ -30,7 +30,7 @@ public class VehicleLineJourney {
     public long getId() {
         return id;
     }
-    public double getActualTime() {
+    public int getActualTime() {
         return actualTime;
     }
     public Vehicles getVehiclesId() {
